feat(iestorybook): make storybook work with ie
