+++ package.json
@@ -100,6 +100,7 @@
     "@react-aria/button": "3.2.2",
     "@react-aria/focus": "^3.4.0",
     "@restart/context": "^2.1.4",
+    "@storybook/addon-ie11": "^0.0.1",
     "@svgr/webpack": "^5.5.0",
     "@types/react": "^17.0.11",
     "classnames": "^2.3.1",
